+++ src/components/TodoList.ts
@@ -1,5 +1,5 @@
-import { setItem } from "../util/storage.js";
-import validation from "../util/validation.js";
+import { setItem } from "../utils/storage.js";
+import validation from "../utils/validation.js";
 
 // params.$target - 해당 컴포넌트가 추가될 DOM 앨리먼트
 // params.initialState - 해당 컴포넌트의 초기 상태
